backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tortoise-orm==0.20.0
pydantic==2.5.0
aiomysql==0.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2